Regroup memory_sync,buffer tests
